Separate contour namespaces out of the base.
So that overlays can more easily specify the deployment namespace,
remove the namespace from the base Contour kustomization. The
workspace that uses this as the base not has to create the deplouyment
namespace.

This also separates out the CRDs into a separate base, but that
doesn't really help many use cases yet.