Deploy knative with contour.
There doesn't seem to be a way to resolve all the corner cases when
trying to import the same base into different namespaces, so we
manually create separate kustomizations for the internal and external
Contour.  Using a custom namespace and nameSuffix seems to work OK.

Add a makefile to tie all the steps together. Add the Docker build
since we have a place to put it.